move validate input to its own file
